finished

+++ model/Task.php
@@ -18,17 +18,18 @@ require_once('../helper/sanitize/Task.php');
     }
     
     
-    public function getTaskById($task_id) {
+    public function getTaskById($task_id, $user_id) {
       
       try {
         
-      $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE id = :id');
+      $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE id = :id AND user_id = :user_id');
       $this->DB->bind(':id', $task_id);
+      $this->DB->bind(':user=id', $user_id);
       $this->DB->execute();
       $_rowCount = $this->DB->rowCount();
  
       if($_rowCount === 0) {
-        $this->response->errorResponse(404, "Task Not Found");
+        $this->response->errorResponse(404, "Task Not Found or Not Authorized");
       } 
 
        // just sanitizing.
@@ -59,10 +60,11 @@ require_once('../helper/sanitize/Task.php');
       } 
     }
 
-    public function deleteById($task_id) {
+    public function deleteById($task_id, $user_id) {
       try {
-        $this->DB->query('DELETE FROM tasks WHERE id = :id');
+        $this->DB->query('DELETE FROM tasks WHERE id = :id AND user_id = :user_id');
         $this->DB->bind(':id', $task_id);
+        $this->DB->bind(':user_id', $user_id);
         $this->DB->execute();
         
         $affectedRows = $this->DB->rowCount();
@@ -71,7 +73,7 @@ require_once('../helper/sanitize/Task.php');
           $this->response = new Response();
           $this->response->setHttpStatusCode(500);
           $this->response->setSuccess(false);
-          $this->response->addMessage("Cound Not find the id specified, thus deletion is not completed.");
+          $this->response->addMessage("Cound Not find Task with Given Id or Not Authorized.");
           $this->response->send();
           // Below this code won't fire anymore
           exit();
@@ -93,10 +95,11 @@ require_once('../helper/sanitize/Task.php');
       }
     }
 
-    public function getTasksByCompletion($_completed) {
+    public function getTasksByCompletion($_completed, $user_id) {
       try {
-        $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE completed = :completed');
+        $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE completed = :completed AND user_id = :user_id');
         $this->DB->bind(':completed', $_completed);
+        $this->DB->bind(':user_id', $user_id);
         $this->DB->execute();
 
         $_rowCount = $this->DB->rowCount();
@@ -135,9 +138,10 @@ require_once('../helper/sanitize/Task.php');
       }
     } 
 
-    public function readPerPage($page, $limitPerPage) {
+    public function readPerPage($page, $limitPerPage, $user_id) {
       try {
-        $this->DB->query('SELECT count(id) as totalNoOfTasks from tasks');
+        $this->DB->query('SELECT count(id) as totalNoOfTasks from tasks WHERE user_id = :user_id');
+        $this->DB->bind('user_id', $user_id);
         $this->DB->execute();
 
         $row = $this->DB->getResult();
@@ -161,8 +165,9 @@ require_once('../helper/sanitize/Task.php');
 
         $offset = ($page == 1 ? 0 : ($limitPerPage * ($page - 1)));
 
-        $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks LIMIT :offset, :pglimit');
+        $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE user_id = :user_id LIMIT :offset, :pglimit');
         $this->DB->bind(':offset', $offset);
+        $this->DB->bind(':user_id', $user_id);
         $this->DB->bind(':pglimit', $limitPerPage);
         $this->DB->execute();
 
@@ -203,12 +208,11 @@ require_once('../helper/sanitize/Task.php');
     }
   }
 
-  public function getAllTasks() {
+  public function getAllTasks($user_id) {
     try {
-      $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks');
+      $this->DB->query('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tasks WHERE user_id = :user_id');
+      $this->DB->bind(':user_id', intval($user_id));
       $this->DB->execute();
-
- 
 
       $_rowCount = $this->DB->rowCount();
 
@@ -247,18 +251,20 @@ require_once('../helper/sanitize/Task.php');
 
   }
 
-  public function createTask($jsonData) {
+  public function createTask($jsonData, $user_id) {
     try {
       
       $newTask = new TaskSanitize(null, $jsonData["title"], isset($jsonData["description"]) ? $jsonData["description"] : null, isset($jsonData["deadline"]) ? $jsonData["deadline"] : null, $jsonData["completed"]);
 
       $sanitizedArray = $newTask->returnTaskAsArray();
 
-      $this->DB->query('INSERT INTO tasks (title, description, deadline, completed) VALUE (:title, :description, STR_TO_DATE(:deadline, "%d/%m/%Y %H:%i"), :completed)');
+      $this->DB->query('INSERT INTO tasks (title, user_id, description, deadline, completed) VALUE (:title, :user_id, :description, STR_TO_DATE(:deadline, "%d/%m/%Y %H:%i"), :completed)');
       $this->DB->bind(":title", $sanitizedArray['title']);
       $this->DB->bind(":description", $sanitizedArray['description']);
       $this->DB->bind(":deadline", $sanitizedArray['deadline']);
       $this->DB->bind(":completed", $sanitizedArray['completed']);
+      $this->DB->bind(":user_id", intval($user_id));
+
       $this->DB->execute();
 
       $insertedId = $this->DB->getLastId();
@@ -306,13 +312,15 @@ require_once('../helper/sanitize/Task.php');
 
     }
 
-    public function updateTask($task_id, $requestData, $updateQuery, $checksheet) {
+    public function updateTask($task_id, $requestData, $updateQuery, $checksheet, $user_id) {
       try {
-        $this->DB->query("SELECT * FROM tasks WHERE id = :id");
+        $this->DB->query("SELECT * FROM tasks WHERE id = :id AND user_id = :user_id");
         $this->DB->bind(':id', $task_id);
+        $this->DB->bind(':user_id', $user_id);
         $this->DB->execute();
+
         if($this->DB->rowCount() == 0) {
-          $this->response->errorResponse(400, "Id is incorrect");
+          $this->response->errorResponse(400, "Id is incorrect or Not Authorized");
         }
        
      
@@ -320,7 +328,10 @@ require_once('../helper/sanitize/Task.php');
         SET 
           $updateQuery
         WHERE 
-          id = :id");
+          id = :id
+        AND
+          user_id = :user_id
+        ");
 
         foreach($checksheet as $key => $value) {
           if($value == true) {
@@ -329,6 +340,7 @@ require_once('../helper/sanitize/Task.php');
         }
 
         $this->DB->bind(":id", $task_id);
+        $this->DB->bind(":user_id", $user_id);
 
         $this->DB->execute();
 
@@ -337,8 +349,9 @@ require_once('../helper/sanitize/Task.php');
         }
 
   
-        $this->DB->query("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed FROM tasks WHERE id = :id");
+        $this->DB->query("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed FROM tasks WHERE id = :id AND user_id = :user_id");
         $this->DB->bind(':id', $task_id);
+        $this->DB->bind(":user_id", $user_id);
         $this->DB->execute();
   
         $numRowAffected = $this->DB->rowCount();
